requirements.txt
alembic==1.16.1
anyio==4.9.0
bcrypt==4.0.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.2
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emails==0.6
fastapi==0.95.2
google-auth==2.17.3
greenlet==3.2.2
h11==0.16.0
idna==3.10
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.7.0
passlib==1.7.4
premailer==3.10.0
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.22.0
PyJWT
jinja2==3.1.2